feat: Build Time Off Form and Refactor Team Dashboard
